requirements.txt
ass==0.5.2
backports.entry-points-selectable==1.1.1
call==0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
data==0.4
decorator==5.1.0
distlib==0.3.3
drop==0.4.1
filelock==3.4.0
Flask==2.0.2
funcsigs==1.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
location==0.0.7
MarkupSafe==2.0.1
platformdirs==2.4.0
print==1.3.0
pymusixmatch==0.3
pyperclip==1.8.2
requests==2.26.0
six==1.16.0
Student==0.0.1
urllib3==1.26.7
virtualenv==20.10.0
Werkzeug==2.0.2
